fix: rm prisma and config split from next dev app
