+++ apps/dev/nextjs/auth.ts
@@ -1,5 +1,8 @@
 import NextAuth from "next-auth"
-import authConfig from "auth.config"
+import type { NextAuthConfig } from "next-auth"
+import Credentials from "next-auth/providers/credentials"
+import Keycloak from "next-auth/providers/keycloak"
+
 // import { PrismaClient } from "@prisma/client"
 // import { PrismaAdapter } from "@auth/prisma-adapter"
 // import SendGrid from "next-auth/providers/sendgrid"
@@ -33,8 +36,45 @@ import authConfig from "auth.config"
 //   }
 // )
 
+declare module "next-auth" {
+  /**
+   * Returned by `useSession`, `getSession`, `auth` and received as a prop on the `SessionProvider` React Context
+   */
+  interface Session {
+    user: {
+      /** The user's postal address. */
+      address: string
+    } & User
+  }
+
+  interface User {
+    foo?: string
+  }
+}
+
 export const { handlers, auth, signIn, signOut, unstable_update } = NextAuth({
-  // adapter: PrismaAdapter(globalThis.prisma),
+  debug: true,
+  providers: [
+    Credentials({
+      credentials: { password: { label: "Password", type: "password" } },
+      authorize(c) {
+        if (c.password !== "password") return null
+        return {
+          id: "test",
+          name: "Test User",
+          email: "test@example.com",
+        }
+      },
+    }),
+    Keycloak,
+  ],
+  callbacks: {
+    jwt({ token, trigger, session }) {
+      if (trigger === "update") token.name = session.user.name
+      return token
+    },
+  },
+  basePath: "/auth",
+  secret: "secret",
   session: { strategy: "jwt" },
-  ...authConfig,
 })
